move artifact utils and fix heaptrack for remote exec
