+++ lib/compilers/llvm-mos.ts
@@ -28,6 +28,7 @@ import type {CompilationResult} from '../../types/compilation/compilation.interf
 import type {PreliminaryCompilerInfo} from '../../types/compiler.interfaces.js';
 import type {ParseFiltersAndOutputOptions} from '../../types/features/filters.interfaces.js';
 import {ArtifactType} from '../../types/tool.interfaces.js';
+import {addArtifactToResult} from '../artifact-utils.js';
 import * as utils from '../utils.js';
 
 import {ClangCompiler} from './clang.js';
@@ -88,7 +89,7 @@ export class LLVMMOSCompiler extends ClangCompiler {
             }
 
             if (await utils.fileExists(nesFile)) {
-                await this.addArtifactToResult(res, nesFile, ArtifactType.nesrom);
+                await addArtifactToResult(res, nesFile, ArtifactType.nesrom);
             }
         } else if (this.compiler.exe.includes('c64')) {
             let prgFile = outputFilename;
@@ -97,7 +98,7 @@ export class LLVMMOSCompiler extends ClangCompiler {
             }
 
             if (await utils.fileExists(prgFile)) {
-                await this.addArtifactToResult(res, prgFile, ArtifactType.c64prg);
+                await addArtifactToResult(res, prgFile, ArtifactType.c64prg);
             }
         }
